fix: 平台 crate 首编清零（CI 编译补盲 job 实证，交叉编译本地复现修复）——linux：dns 解构顺序颠倒致 payload 绑到 bool、fanotify_init 常量 i32→c_uint、aya 0.13 RingBuf 校准（map_mut+try_into）与 Bpf→Ebpf 废弃别名；macos：lsof 刷新线程闭包改持 quad_pid 的 Arc 消生命周期错、sysctl 指针可变性、auditpipe Vec 显式 u8、StartTime 无 Default 改 StartTime(0)、AUE 事件号对齐 OpenBSM 权威值（OPEN_R=72，造文件改 OPEN_WC=77+CREAT=4，原 0x1003_XXXX 越界占位值作废）。仅达 cargo check 级验证，平台功能验证仍属 M2/M3

+++ crates/hg-plat-linux/src/fanotify.rs
@@ -10,7 +10,6 @@
 //! 路径获取：legacy fd 模式（事件携带 fd）→ `readlink /proc/self/fd/<n>`。
 //! 停机：drain 未决权限事件统一 FAN_ALLOW（§9.3，宁放勿卡）。
 
-use std::io::Read;
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicU64, Ordering::Relaxed};
 use std::sync::Arc;
@@ -78,8 +77,12 @@ impl FanotifySource {
         rules: Arc<ArcSwap<RulesSnapshot>>,
         tx: mpsc::Sender<Envelope>,
     ) -> anyhow::Result<Self> {
-        let fd =
-            unsafe { libc::fanotify_init(FAN_CLASS_CONTENT | FAN_CLOEXEC, O_RDONLY_LARGEFILE) };
+        let fd = unsafe {
+            libc::fanotify_init(
+                (FAN_CLASS_CONTENT | FAN_CLOEXEC) as libc::c_uint,
+                O_RDONLY_LARGEFILE as libc::c_uint,
+            )
+        };
         if fd < 0 {
             return Err(anyhow::anyhow!(
                 "fanotify_init 失败 errno={}（需 root）",
